Location page (#107)
* Deleted UserPage and added LocationPage table

* Ran prettier & lint

* Finished Table row creation

* Ran prettier & lint

* Fixed multiple Modal when deleting row

* Finished table edit row functionality

* Ran pp & lint & added propTypes

* Modulized functions in LocationPage

* Ran lint and pp

---------
