+++ client/src/Components/index.jsx
@@ -1,0 +1,5 @@
+import AddLocationModal from './AddLocationModal';
+import EditLocationModal from './EditLocationModal';
+import DeleteModal from './DeleteModal';
+
+export { AddLocationModal, EditLocationModal, DeleteModal };
