+++ client/src/Pages/Locations.jsx
@@ -8,7 +8,7 @@ import {
   flexRender,
 } from '@tanstack/react-table';
 import PropTypes from 'prop-types';
-import { Modal, Button, Form, Container, Row, Col } from 'react-bootstrap';
+import { AddLocationModal, EditLocationModal, DeleteModal } from '../Components';
 
 const columns = [
   {
@@ -65,149 +65,6 @@ const Filters = ({ setColumnFilters }) => {
 
 Filters.propTypes = {
   setColumnFilters: PropTypes.func,
-};
-
-const DeleteModal = ({ toggleDeleteModal, setToggleDeleteModal, row, data, setData }) => {
-  const onDelete = async () => {
-    setToggleDeleteModal(false);
-    try {
-      await fetch(`/api/locations/${row.original.id}`, {
-        method: 'DELETE',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-      });
-      setData(data.filter((d) => d.id !== row.original.id));
-    } catch (e) {
-      console.log(e);
-    }
-  };
-
-  return (
-    <Modal show={toggleDeleteModal} onHide={() => setToggleDeleteModal(false)}>
-      <Modal.Header closeButton>
-        <Modal.Title>Delete Location</Modal.Title>
-      </Modal.Header>
-      <Modal.Body>Are you sure you want to delete this location?</Modal.Body>
-      <Modal.Footer>
-        <Button variant="secondary" onClick={() => setToggleDeleteModal(false)}>
-          No
-        </Button>
-        <Button variant="danger" onClick={onDelete}>
-          Yes
-        </Button>
-      </Modal.Footer>
-    </Modal>
-  );
-};
-
-DeleteModal.propTypes = {
-  toggleDeleteModal: PropTypes.bool,
-  setToggleDeleteModal: PropTypes.func,
-  row: PropTypes.object,
-  data: PropTypes.array,
-  setData: PropTypes.func,
-};
-
-const EditModal = ({ toggleEditModal, setToggleEditModal, data, setData, editData, setEditData }) => {
-  const onChange = (e) => {
-    const newData = { ...editData };
-    newData[e.target.name] = e.target.value;
-    setEditData(newData);
-  };
-
-  const onSubmit = async (e) => {
-    setToggleEditModal(false);
-    e.preventDefault();
-    try {
-      await fetch(`/api/locations/${editData.id}`, {
-        method: 'PATCH',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify(editData),
-      });
-
-      setData(data.map((location) => (location.id == editData.id ? { ...editData } : location)));
-    } catch (err) {
-      console.log(err);
-    }
-  };
-
-  return (
-    <Modal show={toggleEditModal} onHide={() => setToggleEditModal(false)}>
-      <Modal.Header closeButton>
-        <Modal.Title>New Location</Modal.Title>
-      </Modal.Header>
-      <Modal.Body>
-        <Container>
-          <Form>
-            <Row>
-              <Col xs={18} md={12}>
-                <Form.Group controlId="name">
-                  <Form.Label>Location Name</Form.Label>
-                  <Form.Control name="name" autoFocus value={editData.name} onChange={onChange} />
-                </Form.Group>
-              </Col>
-            </Row>
-            <Row>
-              <Col xs={18} md={12}>
-                <Form.Group controlId="address1">
-                  <Form.Label>Address 1</Form.Label>
-                  <Form.Control name="address1" autoFocus value={editData.address1} onChange={onChange} />
-                </Form.Group>
-              </Col>
-            </Row>
-            <Row>
-              <Col xs={18} md={12}>
-                <Form.Group controlId="address2">
-                  <Form.Label>Address 2</Form.Label>
-                  <Form.Control name="address2" autoFocus value={editData.address2} onChange={onChange} />
-                </Form.Group>
-              </Col>
-            </Row>
-            <Row>
-              <Col xs={9} md={6}>
-                <Form.Group controlId="city">
-                  <Form.Label>City</Form.Label>
-                  <Form.Control name="city" autoFocus value={editData.city} onChange={onChange} />
-                </Form.Group>
-              </Col>
-              <Col xs={4.5} md={3}>
-                <Form.Group controlId="state">
-                  <Form.Label>State</Form.Label>
-                  <Form.Control name="state" autoFocus value={editData.state} onChange={onChange} />
-                </Form.Group>
-              </Col>
-              <Col xs={4.5} md={3}>
-                <Form.Group controlId="zipCode">
-                  <Form.Label>Zip Code</Form.Label>
-                  <Form.Control name="zipCode" autoFocus value={editData.zipCode} onChange={onChange} />
-                </Form.Group>
-              </Col>
-            </Row>
-          </Form>
-        </Container>
-      </Modal.Body>
-      <Modal.Footer>
-        <Button variant="secondary" onClick={() => setToggleEditModal(false)}>
-          Close
-        </Button>
-        <Button variant="primary" onClick={onSubmit} type="submit">
-          Submit
-        </Button>
-      </Modal.Footer>
-    </Modal>
-  );
-};
-
-EditModal.propTypes = {
-  toggleEditModal: PropTypes.bool,
-  setToggleEditModal: PropTypes.func,
-  data: PropTypes.array,
-  setData: PropTypes.func,
-  editData: PropTypes.object,
-  setEditData: PropTypes.func,
 };
 
 const LocationTable = ({ table, data, setData }) => {
@@ -286,7 +143,7 @@ const LocationTable = ({ table, data, setData }) => {
           data={data}
           setData={setData}
         />
-        <EditModal
+        <EditLocationModal
           toggleEditModal={toggleEditModal}
           setToggleEditModal={setToggleEditModal}
           row={propRow}
@@ -306,121 +163,12 @@ LocationTable.propTypes = {
   setData: PropTypes.func,
 };
 
-const AddModal = ({ toggleAddModal, setToggleAddModal, data, setData }) => {
-  const [addData, setAddData] = useState({
-    name: '',
-    address1: '',
-    address2: '',
-    city: '',
-    state: '',
-    zipCode: '',
-  });
-
-  const onChange = (e) => {
-    const newData = { ...addData };
-    newData[e.target.name] = e.target.value;
-    setAddData(newData);
-    console.log(addData);
-  };
-
-  const onSubmit = async (e) => {
-    setToggleAddModal(false);
-    e.preventDefault();
-    try {
-      await fetch('/api/locations', {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify(addData),
-      });
-
-      setData([...data, addData]);
-      console.log(data);
-    } catch (err) {
-      console.log(err);
-    }
-  };
-
-  return (
-    <Modal show={toggleAddModal} onHide={() => setToggleAddModal(false)}>
-      <Modal.Header closeButton>
-        <Modal.Title>New Location</Modal.Title>
-      </Modal.Header>
-      <Modal.Body>
-        <Container>
-          <Form>
-            <Row>
-              <Col xs={18} md={12}>
-                <Form.Group controlId="name">
-                  <Form.Label>Location Name</Form.Label>
-                  <Form.Control name="name" autoFocus value={addData.name} onChange={onChange} />
-                </Form.Group>
-              </Col>
-            </Row>
-            <Row>
-              <Col xs={18} md={12}>
-                <Form.Group controlId="address1">
-                  <Form.Label>Address 1</Form.Label>
-                  <Form.Control name="address1" autoFocus value={addData.address1} onChange={onChange} />
-                </Form.Group>
-              </Col>
-            </Row>
-            <Row>
-              <Col xs={18} md={12}>
-                <Form.Group controlId="address2">
-                  <Form.Label>Address 2</Form.Label>
-                  <Form.Control name="address2" autoFocus value={addData.address2} onChange={onChange} />
-                </Form.Group>
-              </Col>
-            </Row>
-            <Row>
-              <Col xs={9} md={6}>
-                <Form.Group controlId="city">
-                  <Form.Label>City</Form.Label>
-                  <Form.Control name="city" autoFocus value={addData.city} onChange={onChange} />
-                </Form.Group>
-              </Col>
-              <Col xs={4.5} md={3}>
-                <Form.Group controlId="state">
-                  <Form.Label>State</Form.Label>
-                  <Form.Control name="state" autoFocus value={addData.state} onChange={onChange} />
-                </Form.Group>
-              </Col>
-              <Col xs={4.5} md={3}>
-                <Form.Group controlId="zipCode">
-                  <Form.Label>Zip Code</Form.Label>
-                  <Form.Control name="zipCode" autoFocus value={addData.zipCode} onChange={onChange} />
-                </Form.Group>
-              </Col>
-            </Row>
-          </Form>
-        </Container>
-      </Modal.Body>
-      <Modal.Footer>
-        <Button variant="secondary" onClick={() => setToggleAddModal(false)}>
-          Close
-        </Button>
-        <Button variant="primary" onClick={onSubmit} type="submit">
-          Submit
-        </Button>
-      </Modal.Footer>
-    </Modal>
-  );
-};
-
-AddModal.propTypes = {
-  toggleAddModal: PropTypes.bool,
-  setToggleAddModal: PropTypes.func,
-  data: PropTypes.array,
-  setData: PropTypes.func,
-};
-
 const Locations = () => {
   const [data, setData] = useState();
   const [columnFilters, setColumnFilters] = useState([]);
   const [toggleAddModal, setToggleAddModal] = useState(false);
-
+  // const page = parseInt(params.get('page') ?? '1', 10);
+  // const [lastPage, setLastPage] = useState(1);
   useEffect(() => {
     fetch('/api/locations')
       .then((res) => res.json())
@@ -446,7 +194,7 @@ const Locations = () => {
         <button type="button" className="btn btn-primary d-flex align-items-center" onClick={() => setToggleAddModal(true)}>
           New <i className="bi bi-plus-lg" />
         </button>
-        <AddModal toggleAddModal={toggleAddModal} setToggleAddModal={setToggleAddModal} data={data} setData={setData} />
+        <AddLocationModal toggleAddModal={toggleAddModal} setToggleAddModal={setToggleAddModal} data={data} setData={setData} />
         <i className="bi bi-person-fill">Locations</i>
         <Filters setColumnFilters={setColumnFilters} />
       </div>
